Add AST checks: command names, os.system calls and platform package
The check_command_names module does three checks on any command that it
finds:

1. It checks if the command is CamelCase. Issues a warning if not.
2. It checks for a common prefix for all commands; this prevents
   cluttering of command names. Issues a warning if multiple
   prefixes are found.
3. It checks if the command ends with the word "Command" and issues
   a warning if not.

The check_os_system_calls module checks if there's any call to
os.system and suggests using subprocess.check_output instead via a
warning.

The check_platform_usage checks for importing the platform package and
suggests to use sublime.platform() and sublime.arch() instead via a
warning.

Failure to parse a pyfile into an AST is considered an error and so
I issue an error in that case.