Update Python requirements to 3.8+
Python 3.4 is absolutely ancient and 3.8 is also EOL by now, but as long
as it still works we may be interested in continuing to use it(since ST
plugins themselves still runs with it).

Required updating github3.py.

Also, Sublime Text 3 is not the latest anymore, so we just remove the
major version specifier.

Disclaimer: I did not check which version the syntax errors have changed
in, so tests could be failing for Python <3.12.